requirements.txt
pandas
numpy
matplotlib
torch
transformers
datasets
scikit-learn
openai
nltk
python-dotenv
google-generativeai